code aufgeraeumt und kommentiert
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@16845 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 8ccac305717cbcd200eef7a067decef0efd20fb5

+++ billingmodel/de.uniol.inf.is.odysseus.billingmodel/src/de/uniol/inf/is/odysseus/billingmodel/IBillingManager.java
@@ -3,21 +3,21 @@ package de.uniol.inf.is.odysseus.billingmodel;
 import java.util.Map;
 
 public interface IBillingManager {
-	public void addPayment(String userID, int queryID, double amount);
+	public void addPayment(String userID, int queryID, long amount);
 
-	public void addPaymentSanction(String userID, int queryID, double amount);
+	public void addPaymentSanction(String userID, int queryID, long amount);
 	
-	public void addRevenue(String userID, int queryID, double amount);
+	public void addRevenue(String userID, int queryID, long amount);
 
-	public void addRevenueSanction(String userID, int queryID, double amount);
+	public void addRevenueSanction(String userID, int queryID, long amount);
 	
-	public Map<String, Map<Integer, Double>> getUnsavedPayments();
+	public Map<String, Map<Integer, Long>> getUnsavedPayments();
 	
-	public Map<String, Map<Integer, Double>> getUnsavedPaymentSanctions();
+	public Map<String, Map<Integer, Long>> getUnsavedPaymentSanctions();
 
-	public Map<String, Map<Integer, Double>> getUnsavedRevenues();
+	public Map<String, Map<Integer, Long>> getUnsavedRevenues();
 
-	public Map<String, Map<Integer, Double>> getUnsavedRevenueSanctions();
+	public Map<String, Map<Integer, Long>> getUnsavedRevenueSanctions();
 
 	public int getNumberOfUnsavedPayments();
 	
@@ -29,5 +29,8 @@ public interface IBillingManager {
 	
 	public long getLastTimestampOfPersistence();
 	
+	/**
+	 * persists the temporary (within one persistence interval) stored billing informations
+	 */
 	public void persistBillingInformations();
 }
